+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add command `fresh_window_id` (@BastienDuplessier, reported by @Zer0xxxx)
 - Add command `picture_z` (@xvw, reported by @Zer0xxxx)
 - Add command `spritesheet_z` (@xvw, reported by @Zer0xxxx)
+- Add `purge` parameter for `text_progressive` (@xvw, reported by @Princesseuh)
 
 ### Bug fixes
 - Fix Game_End background that wasn't disposed on `skip_title` activated (@BastienDuplessier, reported by @Zer0xxxx)